api-gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pyjwt==2.8.0
httpx==0.27.0
